+++ app/src/background.js
@@ -167,6 +167,7 @@ async function untranslateCurrentVideo() {
     originalNodeSelector,
     () => document.location.href,
     "div",
+    false,
     true,
   );
 }
@@ -255,6 +256,7 @@ async function untranslateCurrentMobileFeaturedVideoChannel() {
     originalNodeSelector,
     (el) => el.closest("a").href,
     "span",
+    false,
     true,
   );
 }
@@ -265,6 +267,7 @@ async function untranslateCurrentMobileFeaturedVideoChannel() {
  * @param {string} originalNodeSelector
  * @param {Function} getUrl
  * @param {string} createElementTag
+ * @param {boolean} requirePlayer
  * @param {boolean} shouldSetDocumentTitle
  * @returns
  */
@@ -273,9 +276,11 @@ async function createOrUpdateUntranslatedFakeNode(
   originalNodeSelector,
   getUrl,
   createElementTag,
+  requirePlayer = true,
   shouldSetDocumentTitle = false,
 ) {
   if (
+    !requirePlayer ||
     window.YoutubeAntiTranslate.getFirstVisible(
       document.querySelectorAll(
         window.YoutubeAntiTranslate.getPlayerSelector(),
@@ -432,6 +437,18 @@ async function untranslateOtherVideos(intersectElements = null) {
           return;
         }
 
+        // Check if current widget is a playlist, not video
+        if (
+          video.querySelector('a[href*="/playlist?"]') ||
+          window.YoutubeAntiTranslate.getFirstVisible(
+            video.querySelector(
+              "yt-collection-thumbnail-view-model, .media-item-thumbnail-container.stacked",
+            ),
+          )
+        ) {
+          return;
+        }
+
         // Find link and title elements typical for standard videos
         let linkElement =
           video.querySelector("a#video-title-link") ||
@@ -480,17 +497,6 @@ async function untranslateOtherVideos(intersectElements = null) {
 
         // Ignore advertisement video
         if (window.YoutubeAntiTranslate.isAdvertisementHref(linkElement.href)) {
-          return;
-        }
-
-        // Check if current widget is a playlist, not video
-        if (
-          // Playlists include a "list=" parameter in their href
-          linkElement.href.includes("list=") &&
-          // Playlist do not include an "index=" parameter in their href
-          // Only videos in playlists include an "index=" parameter
-          !linkElement.href.includes("index=")
-        ) {
           return;
         }